Change get_physical_core_numbers output to usize

+++ src/linux/processor.rs
@@ -261,7 +261,7 @@ pub fn get_cpu_frequency(cpu_core_index: usize) -> u64 {
         .unwrap_or_default()
 }
 
-pub fn get_physical_core_numbers() -> u64 {
+pub fn get_physical_core_numbers() -> usize {
     let mut s = String::new();
     if File::open("/proc/cpuinfo")
         .and_then(|mut f| f.read_to_string(&mut s))
@@ -286,7 +286,7 @@ pub fn get_physical_core_numbers() -> u64 {
         }
     }
 
-    core_ids_and_physical_ids.len() as u64
+    core_ids_and_physical_ids.len()
 }
 
 /// Returns the brand/vendor string for the first CPU (which should be the same for all CPUs).
